AWS SDK v3 upgrade
