+++ package.json
@@ -27,12 +27,14 @@
     "prepare": "husky install"
   },
   "dependencies": {
+    "@aws-sdk/client-dynamodb": "^3.435.0",
+    "@aws-sdk/lib-dynamodb": "^3.435.0",
     "@dvsa/mes-microservice-common": "1.0.2",
-    "aws-sdk": "^2.1478.0",
     "compare-versions": "^6.1.0",
     "lodash": "^4.17.21"
   },
   "devDependencies": {
+    "@aws-sdk/credential-providers": "^3.435.0",
     "@dvsa/mes-config-schema": "1.5.0",
     "@types/aws-lambda": "^8.10.125",
     "@types/jasmine": "^5.1.1",
