+++ src/functions/getConfiguration/application/__tests__/removeTrailingSlash.spec.ts
