+++ src/functions/getConfiguration/application/__tests__/validateAppVersion.spec.ts
