+++ src/functions/getConfiguration/application/removeTrailingSlash.ts
