+++ src/functions/getConfiguration/application/validateAppVersion.ts
