+++ src/functions/getConfiguration/constants/errors.constants.ts
