+++ src/functions/getConfiguration/domain/config-builder.ts
@@ -1,6 +1,6 @@
 import { RemoteConfig, TestPermissionPeriod } from '@dvsa/mes-config-schema/remote-config';
 import { config } from './config';
-import { getTestPermissionPeriods } from '../framework/test-permission-repository';
+import { getTestPermissionPeriods } from '../framework/aws/test-permission-repository';
 import { warn } from '@dvsa/mes-microservice-common/application/utils/logger';
 import { ExaminerRole } from '../constants/ExaminerRole';
