+++ src/functions/getConfiguration/domain/getBaseApiUrl.ts
@@ -1,4 +1,4 @@
-import { removeTrailingSlash } from './removeTrailingSlash';
+import { removeTrailingSlash } from '../application/removeTrailingSlash';
 
 export const getBaseApiUrl = (): string => {
   const url = process.env.BASE_API_URL || 'https://dev.mes.dev-dvsacloud.uk/v1';
